+++ Sandbox/Examples/Testing/ExampleMenu.h
@@ -5,7 +5,7 @@ class ExampleMenu: public Layer
 public:
 	ExampleMenu();
 	void OnCreate()override;
-	void OnUpdate(float DeltaTime)override;
+	void OnUpdate(double DeltaTime)override;
 	void OnDestroy()override;
 	void OnGUI()override;
 	void OnRender(double deltime)override;
