+++ Sandbox/Examples/Testing/SerializationTesting.h
@@ -6,7 +6,7 @@ class SerializationTestingLayer :public Layer {
 public:
     SerializationTestingLayer();
     void OnCreate()override;
-    void OnUpdate(float deltatime)override;
+    void OnUpdate(double deltatime)override;
     void OnGUI()override;
     void OnDestroy()override;
 	void OnRender(double deltime)override;
